WFLY-14988 Inside clustering remove unnecessary unboxing to clean code

+++ clustering/common/src/main/java/org/jboss/as/clustering/controller/AddStepHandler.java
@@ -141,7 +141,7 @@ public class AddStepHandler extends AbstractAddStepHandler implements Registrati
         if (index == null) {
             context.addResource(PathAddress.EMPTY_ADDRESS, resource);
         } else {
-            context.addResource(PathAddress.EMPTY_ADDRESS, index.intValue(), resource);
+            context.addResource(PathAddress.EMPTY_ADDRESS, index, resource);
         }
         return resource;
     }
